Update OpenGL section.
The OpenGL repos have had an overhaul earlier this year. This update is to reflect the changes.
Update descriptions of go-gl/gl and go-gl/glfw repos.
Remove glow. It is an internal generator tool used to generate the go-gl/gl bindings; most users will not need to know about it unless they are interested in generating custom bindings or fixing issues. But in that case they can find out about glow from the go-gl/gl repository, since it's described in the README.