Cleanup Web Frameworks section
The Web Framework section was becoming littered with libraries, routers etc. that should not be considered "Web Frameworks",
to that end some libraries have been shuffled to their appropriate sections and a new "Routers" sub-section was created to
distinguish url routers and multiplexers.

I propose that some guidelines or a checklist should be created to help determine what should be considered a "Web Framework"
vs Something Else to avoid the section becoming cluttered again; but that can be done at a later date.